Add tests for gameServerCacheEntry in GameServerAllocation controller
This moves the cache into its own file to make things a bit more
readable, and also to implement some tests for the struct.
