+++ pkg/gameserverallocations/controller.go
@@ -20,7 +20,6 @@ import (
 	"math/rand"
 	"net/http"
 	"strconv"
-	"sync"
 	"time"
 
 	"agones.dev/agones/pkg/apis"
@@ -76,57 +75,6 @@ type Controller struct {
 	stop                <-chan struct{}
 	workerqueue         *workerqueue.WorkerQueue
 	recorder            record.EventRecorder
-}
-
-// gameserver cache to keep the Ready state gameserver.
-type gameServerCacheEntry struct {
-	mu    sync.RWMutex
-	cache map[string]*stablev1alpha1.GameServer
-}
-
-// Store saves the data in the cache.
-func (e *gameServerCacheEntry) Store(key string, gs *stablev1alpha1.GameServer) {
-	e.mu.Lock()
-	defer e.mu.Unlock()
-	if e.cache == nil {
-		e.cache = map[string]*stablev1alpha1.GameServer{}
-	}
-	e.cache[key] = gs.DeepCopy()
-}
-
-// Delete deletes the data. If it exists returns true.
-func (e *gameServerCacheEntry) Delete(key string) bool {
-	e.mu.Lock()
-	defer e.mu.Unlock()
-	ret := false
-	if e.cache != nil {
-		if _, ok := e.cache[key]; ok {
-			delete(e.cache, key)
-			ret = true
-		}
-	}
-
-	return ret
-}
-
-// Load returns the data from cache. It return true if the value exists in the cache
-func (e *gameServerCacheEntry) Load(key string) (*stablev1alpha1.GameServer, bool) {
-	e.mu.RLock()
-	defer e.mu.RUnlock()
-	val, ok := e.cache[key]
-
-	return val, ok
-}
-
-// Range extracts data from the cache based on provided function f.
-func (e *gameServerCacheEntry) Range(f func(key string, gs *stablev1alpha1.GameServer) bool) {
-	e.mu.RLock()
-	defer e.mu.RUnlock()
-	for k, v := range e.cache {
-		if !f(k, v) {
-			break
-		}
-	}
 }
 
 // findComparator is a comparator function specifically for the
